Started factoring out the views file.
